requirements.txt
faster-whisper
rich
sounddevice
soundfile
numpy
python-dotenv
gdown
openai
nemo_toolkit[asr]
